requirements.txt
amqp==2.5.2
Babel==2.8.0
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
debtcollector==1.22.0
Django==2.2.12
django-common-helpers==0.9.2
django-compat==1.0.15
django-cors-headers==3.2.1
django-countries==6.1.2
django-environ==0.4.5
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
facebook-sdk==3.1.0
Flask==1.1.2
future==0.18.2
geoip2==3.0.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
html5lib==1.0.1
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
inflection==0.3.1
iso8601==0.1.12
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.11.1
keystoneauth1==4.0.0
kombu==4.6.8
MarkupSafe==1.1.1
maxminddb==1.5.2
msgpack==1.0.0
netaddr==0.7.19
netifaces==0.10.9
oauthlib==3.1.0
os-service-types==1.7.0
oslo.config==7.0.0
oslo.i18n==3.25.1
oslo.serialization==2.29.2
oslo.utils==3.42.1
packaging==20.1
pbr==5.4.5
phonenumbers==8.12.2
Pillow==7.1.2
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-keystoneclient==3.22.0
python-swiftclient==3.9.0
python-twitch-client==0.6.0
python-twitter==3.5
pytz==2019.3
PyYAML==5.3.1
redis==3.5.2
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
schedule==0.6.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
twilio==6.38.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.4
zipp==1.2.0
